Subtle: when creating symlinks, do NOT include $(DESTDIR) as part of the source of the link.  Otherwise when installing an MFS with something like DESTDIR=/mnt, you wind up with links like:
lrwxr-xr-x   1 root  wheel   25 Nov  9 11:45 lib@ -> /mnt/usr/local/etc/emulab

which do not resolve when the MFS is booted.

This assumes that DESTDIR is only used for temporary mounting of filesystems
that are normally mounted elsewhere (and I believe that has always been the
intent of DESTDIR, at least in BSD).  If DESTDIR reflected the true path,
then the links will now be wrong.