Checkpoint the client side of ElabInElab changes in case some whacko is dying to see whats going on. These changes can go into new images; it will be ignored in non ElabInElab experiments;
rc.inelab: A terrible hack that I now regret. The whole point of this
script to allow inner boss/ops to 1) play by the standard state machine goo
(ISUP) for the outer emulab. That way os_setup, node_reboot, etc all work
when dealing with setting up the inner elab. 2) Since the inner
control can fall on any of the interfaces (as picked by assign), we
need to still do ifc/route configuration using tmcd. At the time I did
this, I expected that swapmod/swapin/swapout would be possible and
that boss/ops would be coming in on different nodes using different
interfaces for the inner control net, and that I would want to ask
each reboot. Well, I think swapping an inner emulab is a really long
way off, maybe so far off I'll be retired by that time.

rc.mkelab: A lot of very ugly code that turns a node into an inner ops
or boss. I won't bother to describe this code other than to say the
operating model for this rev is to create a fully functional inner
emulab based on the DB state provided by the outer emulab. Currently,
the current testbed software is expected to be in the /proj tree, and
a full boss/ops install is done, followed by all kinds of
customizations (like building accounts for members of the project,
etc) to make it a fully function emulab with just a single project;
the project in which the inner emulab was created.