Comment out the `CREATE_HOME' option.
This is done because we replace ceratin user- and group-management programs
(from the `shadow_utils' package) with versions that accept uppercase letters
in user and group names.  The Linux-standard (?) versions of those utilities do
not understand the CREATE_HOME option in the `/etc/login/defs' file.  In fact,
the standard `useradd' program goes out of its way to *complain* about the
option when it is found.  The relevant source comment from `src/useradd.c':

		/*
		 * RedHat added the CREATE_HOME option in login.defs in their
		 * version of shadow-utils (which makes -m the default, with
		 * new -M option to turn it off). Unfortunately, this
		 * changes the way useradd works (it can be run by scripts
		 * expecting some standard behaviour), compared to other
		 * Unices and other Linux distributions, and also adds a lot
		 * of confusion :-(.
		 * So we now recognize CREATE_HOME and give a warning here
		 * (better than "configuration error ... notify administrator"
		 * errors in every program that reads /etc/login.defs). -MM
		 */

Rather than suffer obnoxious warnings at boot time, we just avoid the option.