When the elabinelab experiment is also firewalled, ssh into the firewall node and disable the rules during the inner elab setup, and then turn them back on after the inner boss has rebooted. In the case that an experiment is to be launched inside, launch the experiment async and then turn rules back on. Technically, this should be proxied through the firewall instead of directly, but this is okay for now.
As for experiment teardown, I am not doing anything yet since the
closed firewall lets ssh through, and thats all I need to teardown the
inner elab.

Also during teardown, if DHCPD cannot be killed on inner boss, then
skip rest of the steps and return okay so that the rest of experiment
teardown proceeds (if need be, inner nodes will be power cycled). Not
being be able to kill DHCPD can happen for lots of reasons (like,
experiment never setup in the first place).