Add little tip tool ballons to each dot, which give the nodeid for all nodes, and then if allocated and person has permission, the vname and pid/eid.
Note that tip tool ballons have no formatting capabilities at all;
just a simple text string.

If we want fancier, I found some javascript code that will popup real
windows, and then kill them like a tool tip after a short pause or
when the user mouses out. The downside of the real window is that with
some window managers, you still get a status bar along the top, and so
it looks icky. The upside is that we can put real information and
formatting directives in the popup.