Added support for a switch being in more than one stack at a time. Each switch has a 'primary' stack that it belongs to if it's specified with the '-i' parameter. Otherwise, it can be considered to be a part of any of the stacks of which it's a member.
The main point of this is so that we can have switches that are on
both the control and experimental networks.

Note: Having a VLAN with the same name on two overlapping stacks is
like crossing the streams: that would be bad. Not "all life as you
know it stopping instantaneously" bad, but snmpit might get confused.