Add support for the set-type-limit command in the ptop file.
It looks like this:
set-type-limit <type> <count>

This can be placed anywhere in the ptop file.  It limits the number of
physical nodes of a given type that an experimenter is allowed to use.

For example, we may limit a user to 10 PCs, or 5 pc850s. In the latter
case, if the experimenter had simply asked for 10 generic PCs, then we
will try to find them some mapping of nodes that uses types of PCs
other than pc850s.

Added a precheck to tell the user up-front if they asked for more
nodes than they are allowed to use of a given type.