Add a Samba template config file.  The listening interface, globally allowed access list, and workgroup are defined as follows:
   workgroup = @OURDOMAIN@
   hosts allow = @CONTROL_NETWORK@/@CONTROL_NETMASK@
   interfaces = @FSNODE_IP@

I believe this is what we want, but am not sure about the workgroup setting.
Note that this should ultimately just be the head of the smb.conf file, with
the tail generated vis-a-vis /etc/exports .  For now it has some hard wired
shares for /proj and /users, which is OK for the very short term on an
inner elab.