The Robot Lab Monitor Daemon. A very silly script that looks at some sitevars to determine if the Robot Lab is open or closed. The sitevars:
* 'robotlab/override' - Override other settings and forcibly turn the lab
  "on" or "off" (open or close). When the lab is turned off, new
  experiments cannot swap in and the current experiment is immediately
  swapped out.

* 'robotlab/exclusive' - The robot lab is exclusive use. Best to not mess
  with this sitevar :-)

* 'robotlab/opentime' - The time that the robot lab opens in the
  morning. The default is 07:00, but feel free to change this as you like.

* 'robotlab/closetime' - The time that the robot lab closes in the
  evening. The default is 18:00, but feel free to change this as you like.

* 'robotlab/open' - The robot lab is open or closed. DO NOT MESS WITH THIS!
  It is updated by the robomonitord script and intended to be used by
  admission control (not done yet).

The robomonitord script runs and periodically (every 2 minutes) wakes up
and looks at the various sitevars above. The lab is open during the day,
Monday through Friday, and closed on weekends. It is also supposed to be
closed on holidays, but I have not added that yet.

15 minutes before the lab is to be closed, a warning message is sent to the
swapper of the experiment running on the robot testbed, that their
experiment is going to be swapped soon. When the Robot lab is closed
(either cause the close time was reached, or because the lab was forcibly
closed with the override), the current experiment is immediately swapped
out.

I know, this is hopelessly bogus, but it will do until we feel like adding
a "Lab" datatype to the system.