Stargate client-side boot support.
Similar to linux, but the sg linux distribution is based on debian, so some
things are a little different.

WARNING: if you do a client-install in a cross-compile environment
(with DESTDIR set), watch out for bogus symlinks in <root>/etc/rc*.d!
Have to fix these up by hand in this case.

Also note that the stargate rc script doesn't yet start the robot pilot
daemon - Tim said he needed to do finalize some stuff w.r.t. this, so
he would take care of modifying the rc script.

The ntp settings are still not ideal.  Even though the root fs is on a
journalling jffs filesys, its still not good to write to it regularly.  Need
to point the ntp drift file off into /var (ramfs)