Add tb-set-delay-capacity NS directive:
	tb-set-delay-capacity 1

Will override the default delay capacity as specfied in the node_types
table for each node type, and set it for all types when generating the
ptop file.

This is a big stick, to be used with caution, as it will effectively
double the number of nodes used for delay nodes (withing an experiment).