Support for reloading garcias.
Currently this is a total hack; we simply rsync the stargate as it is freed
from the experiment.  However, until we unravel os_load dependancies between
nodes and subnodes (motes and stargates in this case), doing this through
the appropriate setup channels won't work.

The tbrsync script borrows much of its infrastructure from Rob's tbuisp
script.