Add support for uploading TinyOS images. This page now accepts a 'nodetype=mote' parameter to display only the options that matter for a mote.
This is implemented much like the menu and the beginexp page, with a
$view hash, set from the nodetype argument, that controls
showing/hiding different parts of the page.

Added a file uplaod box to be used for mote images, which calls Tim's
webcopy script to copy the file into place.

Doesn't work quite yet - for some reason, even though the call to
webcopy is succeeding, the destination file doesn't show up. So we
have a little bit of debugging to do still.

Not linked in from anywhere yet. Not sure what the right thing is -
shoud there be a list at the top of this page of all special nodetypes
it supports? Or should we just link it in from the TinyOS tutorial or
the mote type pages?