New web cameras are a lot easier to deal with! I've removed the grabwebcams startup since it is no longer needed.
The new version of the webcam page connects directly to the cameras
using a URL stored in the DB (webcams table). The URL includes the
user (elabman) and the password, which is fine; not likely to leak
out the password and even if we did, no biggie since its on the private
network and no one can get to it anyway.

I have reduced the frame rate to 5fps (compression 50%) as specified
in the URL, and even thats a lot of bandwidth.

I will work on providing mpeg feeds later!