Added two new options, needed for WhOL support.
The new -b option gives you a string with the current status of a
port (ie. if it's enabled, and which VLAN it's in). After getting
this string, you can then modify the port (change its VLAN, enable it,
whatever). Then, when you're done, you can pass the string -b gave you
back into snmpit with -B, and it should restore the original status
of the port.

Make sure to put the status string in single quotes when you pass it
on the command line. It contains some stuff that the shell won't like,
such as semicolons.