Really important stuff:
Unified the 'dot' handling (status printing) of frisbee and imagezip.
They now both report the number of chunks remaining along with the dots.
Also put out a periodic splat for every GB of uncompressed data we write.
This is useful when you are zero-filling, since otherwise it appears that
frisbee has hung when it is really just zeroing the last unused 100GB of
your disk.