Add whack-on-lan power cycle module.
To enable WhOL, you need to add outlets table entries for nodes which are
whol-enabled.  The power_id encodes the interface on boss to use:

+---------+-----------+--------+----------------+
| node_id | power_id  | outlet | last_power     |
+---------+-----------+--------+----------------+
| pcwf6   | whol-fxp0 |      0 | 20050318152119 |
+---------+-----------+--------+----------------+

You then need interfaces and wires table entries for that interface on
boss, so that snmpit works with syntax like "boss:1".  This is probably not
really needed once the VLAN has been setup.

You need a magic VLAN called "WhOL", I used VLAN 999.  Add it to all the
switches and trunks.  Put boss's port in it.  It will remain there, enabled,
forever.

In the interfaces table entry for every interface that supports WhOL,
you need to set the 'whol' field to 1.