Aside from some cosmetic changes and some bug fixes.
The main change is that I now add an .htpasswd entry for the user's
emulab ID (stoller) in addition to the wikiname (LeighStoller). I also
create a mapping file (with perl dbmopen) to map from the Emulab ID to
the WikiName. This lets people log in using their Emulab ID, which is
pretty much required since we cannot pass the token to http basic auth
when it puts up the dialog box. Without this, we will forever be
answering questions from confused users.

TWiki already had something of a mapping mechanism, but it was broken
and horribly inefficient, so I ended up whacking the twiki code to
make it work with the above mentioned dbm file.