Add support for .exe binary files - we process them into the srec files that are given to TinyOS.
As part of this, support changing the ID programmed into the mote.
The user can supply this on the command line, or they can supply it
in the DB (soon to be hooked in to NS files), or, if neither of
these is given, it defaults to the numeric part of the node's name.

This is neat - it could be the start of a more general framework,
in which we do things like modify uploaded TinyOS kernels to use
the frequency assigned to the experiment, or let users set parameters
for their TinyOS kernels in their NS fields.