Pump up grabwebcam script. Actually, I threw out the existing code that picked up single images with scp, and replaced it with stuff that can be used to capture movies.
	Usage: grabwebcams [-d] [-v] [-m] [-t timeout]
	       grabwebcams [-d] [-v] [-t timeout] [-k] pid eid
	switches and arguments:
	-d      - Debug mode, use to prevent daemonization
	-v      - Verbose mode (causes vlc to spit lots of goo)
	-m      - Movie option; create a 10fps movie from each camera
	-t <N>  - Terminate automatically and N seconds
	-k      - Kill a daemonized grabwebcams (only use with pid/eid)
	pid eid - Project and Experiment (for use with swapin)

The first form allows you to capture a low frame rate movie from the
webcams. Low means 2fps. The files are written in the current
directory. Without the -d option, the script goes into the background
and runs forever, unless you give it a -t option, in which case the
movies will be terminated after that many seconds.

Add the -m option, and you get 15fps movies, suitable for showing off
to people. This option IS VERY CPU INTENSIVE. Use it sparingly. It
might get better once I hear back from the axis people, telling me if
there is a way to turn up the mpeg fps via a URL (like most everything
else can be on these cameras). Until then, I am transcoding mjpeg into
mpeg, and that sucks.

The second form is for use from the swapexp path, to start capturing
movies when an experiment is swapped in, and then to then to kill it
off (-k) when the experiment is swapped out. At 2fps, it generates
about 1MB per camera per minute. Thats a lot of data, so I am not
hooking this in quite yet; want to give it some more thought.

We may want to export an interface to the web (and commandline) to
that people can take movies of their experiments at particular times.
Not sure yet.