Hack multicast "keep alive" mechanism.  The "-K <seconds>" option can be used to force the server to send an IGMP report if it doesn't receive any packets within <seconds> seconds.  As long as the server is receiving packets, it won't send the report.
What I'm not lovin here, is that to send a report I have to drop membership
in the group (socket opt IP_DROP_MEMBERSHIP) and rejoin (IP_ADD_MEMBERSHIP).
Simply trying to do an add membership doesn't work because the kernel thinks
you are already in the group and errs out.  I'm hoping all the up and down
activity doesn't make the switch behave any worse than it already does.