New config variables:
     $emulabconfig{"JAILIPBASE"} = "172.16.0.0";
     $emulabconfig{"JAILIPMASK"} = "255.240.0.0";

These are used to establish a route on boss and ops so they can talk to
inner vnodes.

     $emulabconfig{"MFSTARBALL"} = "tftpboot-elabinelab.tar.gz";
     $emulabconfig{"MFSVERSION"} = "53";
     $emulabconfig{"MFSCONSOLE"} = "sio";

Allow for customization of the MFSes.  The first is the most useful,
it provides some backend support for something Leigh suggested: the ability
to select in the NS file whether the inner-elab should use a "release"
set of files (MFSes, images, emulab source) or the current elabinelab version.
The last two might not be as useful.

Currently none of these new variables are actually passed in via tmcd,
they just get the default values shown above.