Have the locpiper save dynamic obstacles (internally) so that when a new client connects, we can dump the current set to the applet. I'm still holding off on putting them into the DB (really do not think this is the right thing to do, but I can be convinced otherwise).
Change battery code in the locpiper so that if the data for a robot is
more then 10 minutes old, stop sending updates for that robot.

For the applet, add a little daemon thread that checks to see if
battery updates have stopped, and if so, change the color of the robot
to blue (cold and stale). This makes it more clear which robots are
working and which are turned off.

Fix collision detection code so that we test for the center of a robot
overlapping with an obstacle, rather then radius of the area it sweeps
(Tim said this was the correct thing to do).