Add a "nosetup" option to elabinelab experiments. In the experiments table, if elabinelab_nosetup is non-zero, boss and ops setup will do just enough to get the nodes into a state that hopefully approximates what a real installation might look like before installing our stuff. I do install the packages cause there is no point in waiting for that to finish interactively.
From this point, you can log into the console(s) and run the setup
instructions verbatim, although I have not actually tried that yet.

The nice thing is that if you manage to get things properly setup, it
can function as a real elabinelab since the outer environment has been
setup. This is quite a bit different then how we tested during the
last release frenzy.

Its not quite perfect of course, since the images are not "clean", but
I think this is okay for testing.