In addition to the giant commit log file (which is really getting big), generate month by month files, and then a little index to include into the php script.
I know, this is really important stuff! But, I was reminded to do this by a
message that Christopher Johnson (the young one) sent on April 29th about
his project. His code was quaint, but I did not think we needed a new
python script when it was much easier just to roll it into the existing
perl script that imports the commitlog.