Pass around pid,eid (through the applet) so that the backend knows what locpiper to connect to. Locpiper now binds a dynamic port, and stores that into the DB, and the php script expects to get that pid,eid back.
Changed the showexp page so that pid,eid are passed to the robotrack
top level page.

Changed the robotrack page so that if it does not get a pid,eid, it looks
for an experiment running (a locpiper). If none is running, let the applet
start of course, but there will be an error dialog when it starts. You can
still play with the applet of course. If there is more then one experiment
running, then just pick first one. At some point I might add a page to ask
which experiment, but its low priority.