Terrible hack.
Non-modular switches with two different types of interfaces give us
trouble under IOS because they appear to be on the same module, but
the duplicate card numbers. ie. Fastethernet0/1 and Gigabitethernet0/1
Hack around this by pretending the gigabit ports on such a switch are
on module 1, not module 0.

This works fine on the 29xx series, but probably won't work well for
anything else. The real solution is to move to iface instead of
card.port, or use a type.card.port triple.

In order to invoke this hack, you must put '-modhack' in the type for
the switch in the nodes table.