Add a PRUNE_VLANS option, which prevents us from creating VLANs on switches that don't need to have them.
We always create VLANs on the leader, though, so that it can be used
as a sort of global list of what VLANs are available, and is
atomically lockable.

Makes some changes to the private VLAN code which are not yet well
tested - but that's okay, because no one uses private VLANs right now.