Checkpoint CVS repo support stuff. I have added pserver support stuff, to generate the cvsd.conf file from a head+tail file, and restart the server.
A few other steps need to be performed though since we NFS local mount
the project CVS repos into a jail that cvsd is running in. So, the
mount directory has to be changed, and the repo read-only mounted into
the jail. Also setup anoncvs (and anonymous) user entries (no
password), and munge the config file in CVSROOT so that the lockdir
points to someplace in the jail that can be written.