Inner-elab role changes:
	'boss' -> 'boss+router'
	'ops'  -> 'ops+fs'
	'fs'      (new role)
	'router'  (new role)

These are in preparation for allowing configurations with split ops and fs
nodes (and sometime down the road, split boss and router nodes).

This checkin is just the DB state changes along with the scripts that look
at that state.  The Big One, which actually sets up separate nodes
automatically, is undergoing more testing but will be Coming Soon.