Minor interface changes for Jay.
Major interface change; the default mode in the map is that click
on a node now *adds* to a highlight (used to be shift-click), and that
click over the floor area deselects everything.

Add a "Distances between highlighted node" menu item, that uses the
x,y,z coords of nodes to determine the distance between them. Okay
so there might be a small constant in there, say 14 feet, that just
happens to be the distance between floors in MEB.