Move all modification of the group_membership table to the backend, into a single new script call modgroups. Usage:
	modgroups [-a pid:gid:trust[,pid:gid:trust]...]
                  [-m pid:gid:trust[,pid:gid:trust]...]
                  [-r pid:gid[,pid:gid]...] user

So, -a to add groups, -r to remove groups, and -m to modify the trust
value for a member of a group.

The reason for doing this is that previously, we had no idea in the
backend what group changes actually happened; we just knew what the
current groups are. This make it hard to add and remove users from
mailing lists, chat server buddy lists, etc. This is cleaner ...