Add some hacky stuff to allow people in NSDI temporary project to see experiments of other people in the NSDI project.
Added sitevar "general/open_showexplist" which when set to the name of
a project (NSDI), will allow users in that project to view all the
experiments that are swapped in by other members of the project, even
if the experiments are within another project.

Also quickly hacked showuser page to only show project/group listing
of overlapping projects, not all projects. This avoids some leakage.