Add a new option, -N, which prints out the name of the VLAN number assoicated with the named VLAN..
Output format:
<vlan_name> <stack> <vlan_number>
(This is so the user can, if they want, look for more than one
VLAN on more than one stack at a time.)

This should make some of the control-net firewall code cleaner.

For now, though, it's pretty inefficient, becuase it calls listVlans()
on the whole stack, which is much more than needed. But, a more efficient
implementation will require a change to the switch/stack API.