Initial cross domain login support for the datapository. When the remote site (currently just www.datapository.net) wants to verify the user (log the user in), it redirects the user to:
	https://www.emulab.net/login_redirect.php?redirect_to=http://www.datapository.net/login.rhtml

Emulab verifies the user, and then issues an ssh to the host in the
URL, giving it the uid and a freshly generated hash string. Emulab
then redirects the users browser back over to the redirect_to URL,
appending the user name and the hash.

	http://www.datapository.net/login.rhtml?user=$uid&auth='$hash'

What the remote site does with all of this is up to that site.