Have the exp. proc. return better exit codes.  Store the exit codes in a file for easy refrence after the experient is done.
Fork a child to do the real work.  This child will call setsid() and
redirect /dev/null to STDIN so that SSH has absolitly no way to get a
response from the user and will thus give up when public key auth
doesn't work.

Improve logging of experments.

Add test to see if the proj directory is corretly mounted and writable.

Add soft and hard limits to the maximum number of nodes to use.  The
soft has teh same effect to the value given to "-p" is earlier versions.
"-p" no longer takes and parameters.

Create a special script to ssh to the nodes for each experient to avoid
hideously long command lines.

Other misc. changes.