Fix a bug that Mike reported wrt the swap uid being set too late, causing the program agent to run as last swapper.
This was a little tricky cause of a poor decision to share the
usage of the last_swap_uid in the stats gathering code, which wants
the to set the last swapper late so that the previous swapper gets
charged appropriately. Rather then mess with the stats code too much,
I moved things around a bit, setting the swapper earlier and adding
code in libdb to capture the original swapper at the begining of a
swapmod for accounting, and then adding code in swapexp to reset the
swapper if a swapmod fails.

Should fix the stats code at some point to have its own idea of
swapper.