The beginnings of a toy version of assign written in Comet. I'm working on getting a feeling of whether or not it's feasible to implement assign in Comet. I think if it is, it would be a huge win in terms of maintainability and extensibility.
So far, it's looking very promising. I'm struggling a bit with the
syntax, and the lack of an API reference. But, once I figure out how
to do what I need, it's making a lot of sense.

All this program does so far is try to respect the number of slots on
physical nodes - ie. not assign more than one vnode to a pnode, but it
can do arbitrary numbers instead of just 1.

It does not read top/ptop files, doesn't know about types, doesn't try
to minimize any score, doesn't do features/desires, doesn't know about
links at all, etc. Like I said, it's the beginnings of a toy program.

It looks to me like what we're going to end up doing is writing a
program to produce Comet programs. One of the cool things about this
is that it means we can leave out any constraints, etc. that are not
relevant to the particular mapping problem at hand. And, it should
make it easy for us to plug in other search techniques (ie. write the
obvious ones for certain pathological cases, like snakes and
trees.)